Moved image loading to controller
