+++ web/src/themes/themeService.ts
@@ -2,68 +2,9 @@ import { CellView, CellType } from "../viewModel/board/model";
 import { PieceKind } from "../api/model";
 import { Dispatch } from "redux";
 import * as StoreDisplay from '../store/display';
-import ThemeFactory from "./themeFactory";
-import LocalStorageService from "../utilities/localStorageService";
 import { Theme } from './model';
-import { CustomAction } from "../store/root";
 
 export default class ThemeService {
-
-    //#region Theme initialization
-
-    public static changeTheme(themeName : string, dispatch : (action : CustomAction) => void) : void {
-        LocalStorageService.themeName = themeName;
-        const theme = ThemeFactory.getThemes().get(themeName);
-        dispatch(StoreDisplay.Actions.changeTheme(theme));
-        ThemeService.applyToCss(theme);
-        ThemeService.loadThemeImages(theme, dispatch);
-    }
-
-    private static applyToCss(theme : Theme) : void {
-        const s = document.documentElement.style;
-        const c  = theme.colors;
-
-        s.setProperty("--background-color", c.background);
-        s.setProperty("--text-color", c.text);
-        s.setProperty("--header-text-color", c.headerText);
-        s.setProperty("--border-color", c.border);
-        s.setProperty("--hover-text-color", c.hoverText);
-        s.setProperty("--hover-background-color", c.hoverBackground);
-        s.setProperty("--alt-row-background-color", c.altRowBackground);
-        s.setProperty("--alt-row-text-color", c.altRowText);
-
-        s.setProperty("--player-color-0", c.player0);
-        s.setProperty("--player-color-1", c.player1);
-        s.setProperty("--player-color-2", c.player2);
-        s.setProperty("--player-color-3", c.player3);
-        s.setProperty("--player-color-4", c.player4);
-        s.setProperty("--player-color-5", c.player5);
-        s.setProperty("--player-color-6", c.player6);
-        s.setProperty("--player-color-7", c.player7);
-
-        s.setProperty("--font-family", theme.fonts.normalFamily);
-        s.setProperty("--header-font-family", theme.fonts.headerFamily);
-    }
-
-    private static loadThemeImages(theme : Theme, dispatch : (action : CustomAction) => void) : void {
-        const kinds = [
-            PieceKind.Assassin,
-            PieceKind.Chief,
-            PieceKind.Corpse,
-            PieceKind.Diplomat,
-            PieceKind.Gravedigger,
-            PieceKind.Reporter,
-            PieceKind.Thug
-        ];
-        kinds.forEach(k => ThemeService.createPieceImage(theme, k, dispatch));
-    }
-
-    private static createPieceImage(theme : Theme, kind : PieceKind, dispatch : (action : CustomAction) => void) : HTMLImageElement {
-        const image = new (window as any).Image() as HTMLImageElement;
-        image.src = ThemeService.getPieceImagePath(theme, kind);
-        image.onload = () => dispatch(StoreDisplay.Actions.pieceImageLoaded(kind, image));
-        return image;
-    }
 
     public static getPieceImagePath(theme : Theme, kind : PieceKind) : string {
         const i = theme.images.pieces;
@@ -77,17 +18,6 @@ export default class ThemeService {
             case PieceKind.Thug: return i.thug;
             default: throw "Invalid piece kind: " + kind;
         }
-    }
-
-    //#endregion
-
-    public static loadSavedTheme(dispatch : (action : CustomAction) => void) : void {
-        return ThemeService.changeTheme(ThemeService.getSavedThemeName(), dispatch);
-    }
-
-    private static getSavedThemeName() : string {
-        let name = LocalStorageService.themeName;
-        return name ? name : ThemeFactory.default.name;
     }
 
     //#region Colors
